refactor: move frequency DTOs into separate files and directories for better organization
