+++ src/modules/frequency/application/dtos/index.ts
@@ -1,0 +1,3 @@
+export * from "./get-user-classes.dto";
+export * from "./class-attendance.dto"
+export * from "./general-attendance.dto";
